resolve mistakes after review
