+++ src/components/ingredient-details/ingredient-details.module.css
