+++ src/components/modal/modal.jsx
@@ -4,21 +4,26 @@ import PropTypes from "prop-types";
 
 import ModalOverlay from "./modal-overlay.jsx";
 import popupStyles from './popup.module.css'
-import cross from '../../images/cross.svg';
+import { CloseIcon } from "@ya.praktikum/react-developer-burger-ui-components";
 
 const modalRoot = document.getElementById("react-modals");
 
 export default function Modal({ title, onClose, children }) {
+    function onEscKeyDown(e) {
+        if (e.key === "Escape") onClose();
+    }
+
     return createPortal(
         (
         <>
-            <ModalOverlay>
+            <ModalOverlay onClose={onClose} onEscPress={onEscKeyDown}>
                 <div className={`${popupStyles.popup__content} pt-10 pb-15 pl-10 pr-10`} onClick={e => e.stopPropagation()}>
                     <div className={`mb-4 ${popupStyles.popup__base}`}>
                         {title &&
                         <h3 className={`text text_type_main-large ${popupStyles.popup__title}`}>{title}</h3>}
-                        <img onClick={onClose} className={popupStyles.popup__cross} src={cross}
-                             alt="закрытие окна"/>
+                        <div onClick={onClose} className={popupStyles.popup__cross}>
+                            <CloseIcon/>
+                        </div>
                     </div>
                     {children}
                 </div>
