+++ src/components/modal/popup.module.css
@@ -24,12 +24,6 @@
     border: 1px solid rgba(76, 76, 255, 0.2);
 }
 
-.popup__done {
-    width: 120px;
-    height: 120px;
-    margin: 0 auto;
-}
-
 .popup__base {
     display: flex;
     align-items: center;
@@ -41,11 +35,6 @@
     height: 18px;
     margin-left: auto;
     cursor: pointer;
-}
-
-.order__number {
-    text-align: center;
-    text-shadow: 0 0 16px rgba(51, 51, 255, 0.25), 0 0 8px rgba(51, 51, 255, 0.25), 0 4px 32px rgba(51, 51, 255, 0.5);
 }
 
 .popup__text_type_centered {
